# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MachineEmu Rust API",
    "description": "API v2 for instances, live devices, and ticketed streams",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/images/{id}": {
      "get": {
        "tags": [],
        "summary": "Get an image",
        "operationId": "get_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get an image",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [],
        "summary": "Replace an image manifest (JSON or YAML)",
        "operationId": "put_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterImage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replace an image manifest (JSON or YAML)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/instances/{id}/config": {
      "get": {
        "tags": [],
        "summary": "Get saved instance profile and launch plan (Accept: application/yaml supported)",
        "operationId": "get_instance_config",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get saved instance profile and launch plan (Accept: application/yaml supported)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [],
        "summary": "Replace stopped instance profile and launch plan (JSON or YAML)",
        "operationId": "put_instance_config",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replace stopped instance profile and launch plan (JSON or YAML)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v2/profiles/{id}": {
      "get": {
        "tags": [],
        "summary": "Get a shared profile (Accept: application/yaml supported)",
        "operationId": "get_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a shared profile (Accept: application/yaml supported)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [],
        "summary": "Replace a workspace profile override (JSON or YAML)",
        "operationId": "put_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replace a workspace profile override (JSON or YAML)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "HelperSpec": {
        "type": "object",
        "required": [
          "name",
          "argv"
        ],
        "properties": {
          "after_qemu": {
            "type": "boolean"
          },
          "argv": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "ready_socket": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "InstanceConfig": {
        "type": "object",
        "required": [
          "schema_version",
          "instance_id",
          "image_id",
          "profile_id",
          "revision",
          "auto_remove",
          "launch_plan"
        ],
        "properties": {
          "auto_remove": {
            "type": "boolean"
          },
          "image_id": {
            "type": "string"
          },
          "instance_id": {
            "type": "string"
          },
          "launch_plan": {
            "$ref": "#/components/schemas/LaunchSpec"
          },
          "profile": {},
          "profile_id": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "LaunchSpec": {
        "type": "object",
        "required": [
          "argv",
          "qmp_socket"
        ],
        "properties": {
          "argv": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "helper_argv": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "helpers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HelperSpec"
            }
          },
          "preparation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PreparationSpec"
              }
            ]
          },
          "qmp_socket": {
            "type": "string"
          },
          "stderr": {
            "type": [
              "string",
              "null"
            ]
          },
          "stdout": {
            "type": [
              "string",
              "null"
            ]
          },
          "vnc_auto": {
            "type": "boolean"
          }
        }
      },
...
      "PreparationSpec": {
        "type": "object",
        "required": [
          "disk_backing",
          "backing_format"
        ],
        "properties": {
          "backing_format": {
            "type": "string"
          },
          "disk_backing": {
            "type": "string"
          },
          "disk_size": {
            "type": [
              "string",
              "null"
            ]
          },
          "nvram_seed": {
            "type": [
              "string",
              "null"
            ]
          },
          "tpm_seed": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RegisterImage": {
        "type": "object",
        "required": [
          "image_id",
          "engine_track",
          "target",
          "disk_sha256"
        ],
        "properties": {
          "disk_sha256": {
            "type": "string"
          },
          "engine_track": {
            "type": "string"
          },
          "firmware_sha256": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_id": {
            "type": "string"
          },
          "supported_engine_tracks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "target": {
            "type": "string"
          },
          "tpm_state_sha256": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}